Fixes / tested and working
The adjustableValues was getting a copy of the Double instead of the reference. Maybe this could be made to work, but instead I just switched to using a HashMap. Originally I didn't want to use the HashMap because of maybe not being able to control the order of cycling through values, but we can come back to that later if needed.

For gamepad, was copying in the wrong direction. It is a copyFrom but I assumed it was a copyTo.

Added ability to force the motors on or off, irrespective of the analog signal. This logic isn't probably exactly what we want. If the analog signal switches it will take back control.